==> Notebooks/2_data_merge.ipynb <==
## Data Merging

**Name**: Diane Lu

**Contact**: dianengalu@gmail.com

**Date**: 07/31/2023

## Table of Contents 

1. [Introduction](#intro)
2. [Merged Dataset](#merge)
3. [Final Dataset](#final)
4. [Model Dataset](#model)
5. [Vancouver Data](#van)

### Introduction <a class="anchor" id="intro"></a>

In this notebook, our main focus is to merge the cleaned datasets and create the final datasets. We've put in the effort to clean the individual datasets, and now it's time to bring them all together to form the complete and ready-to-use data.

#### Downloading Yelp Dataset from Kaggle 

You can download the necessary Yelp Dataset from [Kaggle](https://www.kaggle.com/datasets/yelp-dataset/yelp-dataset/versions/3?datasetId=10100&sortBy=voteCount&select=Dataset_User_Agreement.pdf). For this data analysis, we will be using Version 3 of the Yelp Dataset found on Kaggle and specifically the business, review, and user datasetS. 

#### Importing Python Libraries 

Importing necessary libraries for the Data Preprocessing.

# Import necessary libraries
import numpy as np 
import pandas as pd 

# Ignore all warnings to avoid cluttering the output
import warnings
warnings.filterwarnings("ignore")

### Merging Datasets for Modeling <a class="anchor" id="merge"></a>

Importing all our cleaned datasets.

# Load the pickled business_data DataFrame
business_data = pd.read_pickle('T:/GitHub/Brainstation_Capstone/Data/business_data.pkl')

# Load the pickled review_data DataFrame
review_data = pd.read_pickle('T:/GitHub/Brainstation_Capstone/Data/review_data.pkl')

# Load the pickled user_data DataFrame
user_data = pd.read_pickle('T:/GitHub/Brainstation_Capstone/Data/user_data.pkl')

Merging `business_data` and `review_data`. 

# Merge the 'review_data' DataFrame and 'business_data' DataFrame based on the common column 'business_id'
business_reviews = review_data.merge(business_data, on='business_id')

# Display the first few rows of the 'business_reviews' DataFrame
business_reviews.head()

Merging `business_reviews` and `user_data`. 

# Merge the 'business_reviews' DataFrame and 'user_data' DataFrame based on the common column 'user_id'
final_business_reviews = business_reviews.merge(user_data, on='user_id')

# Display the first few rows of the 'final_business_reviews' DataFrame
final_business_reviews.head()

# Pickle the DataFrame
final_business_reviews.to_pickle('T:/GitHub/Brainstation_Capstone/Data/final_reviews.pkl')

### Final Dataset for EDA <a class="anchor" id="final"></a>

We've successfully created a final dataset that's all set for exploratory data analysis (EDA) and further manipulation. This dataset is now in excellent shape and ready for us to dive into some in-depth analysis and get valuable insights from it.

**Data Dictionary:**
* `review_id`: unique review id
* `user_id`: unique user id
* `business_id`: unique user id
* `stars`: star rating
* `text`: the review itself
* `restaurant_name`: the restaurant's name
* `address`: the full address of the restaurant
* `city`: the city
* `state`: 2 character state code
* `postal_code`: the postal code
* `latitude`: latitude of the restaurant
* `longitude`: longitude of the restaurant
* `restaurant_rating`: star rating
* `restaurant_review_count`: number of reviews
* `user_review_count`: the number of reviews they've written
* `is_open`: 0 or 1 for closed or open
* `categories`: business categories
* `user_name`: the user's first name
* `average_stars`: average rating of all reviews

final_data = pd.read_pickle('T:/GitHub/Brainstation_Capstone/Data/final_reviews.pkl')

# Display concise information about the 'final_data' DataFrame
final_data.info()

# Display the first few rows of the 'final_data' DataFrame
final_data.head()

final_data.isnull().sum()

### Modeling Dataset <a class="anchor" id="model"></a>

We're going to build a `model_dataset` by setting some thresholds for our data. We'll create one threshold to include only restaurants where the `restaurant_review_count` is greater than 100, and another threshold to include only users where the `user_review_count` is greater than 100. This way, we can focus on a subset of data that meets these criteria and use it to train and build our model effectively.

Creating a threshold where we are only including restaurants where `restaurant_review_count` is greater than 100. 

# Filter 'final_data' DataFrame to include only rows where the 'restaurant_review_count' is greater than or equal to 100
final_data = final_data[final_data['restaurant_review_count'] >= 100]

# Print a sanity check message showing the minimum number of restaurant reviews in the filtered DataFrame
print(f"Sanity Check: The minimum amount of restaurant reviews is {final_data['restaurant_review_count'].min()}.")  

Creating a threshold where we are only including users where `user_review_count` is greater than 100. 

# Filter 'final_data' DataFrame to include only rows where the 'user_review_count' is greater than or equal to 100
final_data = final_data[final_data['user_review_count'] >= 100]

# Print a sanity check message showing the minimum number of user reviews in the filtered DataFrame
print(f"Sanity Check: The minimum amount of user reviews is {final_data['user_review_count'].min()}.")


To encode our ID columns, we're utilizing a label encoder. This encoder will help us convert the categorical ID values into numerical representations, making it easier for our algorithms and models to work with the data.

from sklearn import preprocessing

def label_encode_column(df, column_name, encoder_name):
    # Create an instance of the LabelEncoder
    encoder = preprocessing.LabelEncoder()

    # Use the LabelEncoder to transform the specified column
    df[column_name] = encoder.fit_transform(df[column_name])

    # Store the encoder object in the DataFrame for later use if needed
    df[encoder_name] = encoder

    return df

# Label encode the 'review_id' column in the 'final_data' DataFrame
label_encode_column(final_data, 'review_id', 'review_id_encoder')

# Label encode the 'business_id' column in the 'final_data' DataFrame
label_encode_column(final_data, 'business_id', 'business_id_encoder')

# Label encode the 'user_id' column in the 'final_data' DataFrame
label_encode_column(final_data, 'user_id', 'user_id_encoder')

# Drop the 'review_id_encoder', 'business_id_encoder', and 'user_id_encoder' columns from the 'final_data' DataFrame
final_data = final_data.drop(['review_id_encoder', 'business_id_encoder', 'user_id_encoder'], axis=1)

# Display the first few rows of the 'final_data' DataFrame
final_data.head()

We'll extract specific columns, namely `user_id`, `business_id`, `stars`, `restaurant_name`, and `categories`, from the `final_data` DataFrame. Then, we'll create a new DataFrame called model_`data with these columns.

But, there's one last thing we need to do – let's rename the `stars` column to `rating` in our `model_data` DataFrame. This way, it'll be more intuitive and align with our model's requirements.

# Extract specific columns 'user_id', 'business_id', 'stars', 'restaurant_name', and 'categories' from final_data
model_data = final_data[['user_id', 'business_id', 'stars', 'restaurant_name', 'categories']]

# Rename the 'stars' column to 'rating' in the 'model_data' DataFrame
model_data.rename(columns={'stars': 'rating'}, inplace=True)

# Display the first few rows of the resulting DataFrame
model_data.head()

# Pickle the DataFrame
model_data.to_pickle('T:/GitHub/Brainstation_Capstone/Data/model_data.pkl')

With all our datasets now cleaned and merged, we're good to go for the modeling phase! Our data is in excellent shape and ready for some serious analysis.
