==> Notebooks/4_content_based.ipynb <==
## Content Based Model

**Name**: Diane Lu

**Contact**: dianengalu@gmail.com

**Date**: 07/31/2023

### Table of Contents 

1. [Introduction](#intro)
2. [Model Dataset](#model)

### Introduction <a class="anchor" id="intro"></a>

In this notebook, where we'll be building a content-based recommender system for restaurants. Our goal is to create a system that suggests restaurants to users based on their preferences and past interactions with restaurants. By leveraging the content and features of the restaurants, we can provide personalized recommendations that align with each user's taste. 

#### Importing Python Libraries 

Importing necessary libraries.

# Import necessary libraries
import numpy as np
import pandas as pd

# Import data visualization libraries
import matplotlib.pyplot as plt

# Import NLP (Natural Language Processing) packages
import string
import nltk

# Import TF-IDF (Term Frequency-Inverse Document Frequency) Vectorizer from scikit-learn
from sklearn.feature_extraction.text import TfidfVectorizer

# Import train_test_split function from scikit-learn to split the data into training and test sets
from sklearn.model_selection import train_test_split

# Import cosine_similarity function from scikit-learn to calculate similarity scores
from sklearn.metrics.pairwise import cosine_similarity

# Ignore all warnings to avoid cluttering the output
import warnings
warnings.filterwarnings("ignore")

### Business Dataset <a class="anchor" id="model"></a>

We will begin by importing our Business Dataset and giving it a good inspection to see how it looks. This dataset will serve as a foundational element for our project, so it's crucial to understand its structure and contents before diving into further steps.

**Data Dictionary:**
| Column                   | Description                                           |
|--------------------------|-------------------------------------------------------|
| review_id                | Unique identifier for each review.                   |
| user_id                  | Unique identifier for each user.                     |
| business_id              | Unique identifier for each business (restaurant).    |
| stars                    | Rating given by the user for the restaurant (1 to 5 stars). |
| text                     | The review text provided by the user for the restaurant. |
| restaurant_name          | The name of the restaurant.                          |
| address                  | The address of the restaurant.                       |
| city                     | The city where the restaurant is located.            |
| state                    | The state where the restaurant is located.           |
| postal_code              | The postal code of the restaurant's location.        |
| latitude                 | The latitude coordinate of the restaurant's location.|
| longitude                | The longitude coordinate of the restaurant's location.|
| restaurant_rating        | The overall rating of the restaurant.                |
| restaurant_review_count  | The total number of reviews for the restaurant.      |
| is_open                  | Indicator whether the restaurant is open or closed (1 for open, 0 for closed). |
| categories               | The categories or types of the restaurant (comma-separated strings). |
| name                     | The name of the user who provided the review.        |
| user_review_count        | The total number of reviews provided by the user.    |
| average_stars            | The average rating given by the user to various restaurants. |


# Read data from a pickle file into a Pandas DataFrame
final_reviews = pd.read_pickle('T:/GitHub/Brainstation_Capstone/Data/final_reviews.pkl')

# Display concise information about the 'final_reviews' DataFrame
final_reviews.info()

# Display the first few rows of the 'final_reviews' DataFrame
final_reviews.head()

# Select only the specified columns from the filtered data
final_reviews = final_reviews[['business_id', 'restaurant_name', 'city', 'state', 'restaurant_rating', 'categories']]

# Count the number of missing values in each column of the 'final_reviews' DataFrame
final_reviews.isnull().sum()

# Print the size of our model dataset
print(f"The size of our model dataset is {final_reviews.shape[0]} entries.")

final_reviews = final_reviews[~final_reviews['restaurant_name'].duplicated(keep='first')]

display(final_reviews)

To organize the dataset, we'll sort it by the business ID in ascending order. After that, we'll reassign the IDs so that they start from 0 and continue in ascending order. This way, we'll have a neatly organized dataset with consecutive and easily interpretable business IDs.

# Extract columns 'business_id', 'restaurant_name', and 'rating' from 'final_reviews'
sorted_data = final_reviews.sort_values(by='business_id')

# Display the sorted data
display(sorted_data)

sorted_data['business_id'] = sorted_data['business_id'].rank(method='dense').astype(int) - 1

# Display the updated DataFrame
display(sorted_data)

sorted_data.info()

# Number of restaurants 
print("Number of restaurants:", sorted_data['restaurant_name'].nunique())

# Number of unique reviewers 
print("Number of unique reviewers:", sorted_data['business_id'].nunique())

# Range of ratings
print("Range of ratings:", sorted_data['restaurant_rating'].min(), "to", sorted_data['restaurant_rating'].max())

### Baseline Content Based Filtering <a class="anchor" id="base"></a>

Before applying the TF-IDF (term frequency-inverse document frequency) transformation, we'll be converting the `categories` column from list values to strings. This preprocessing step is necessary to ensure that the data is in the appropriate format for the TF-IDF algorithm. 

sorted_data['categories'] = sorted_data['categories'].apply(lambda x: ', '.join(x))

sorted_data['features'] = sorted_data['categories'] + ', ' + sorted_data['city'] + ', ' + sorted_data['state']

sorted_data.drop(['city', 'state', 'restaurant_rating', 'categories'], axis=1, inplace=True)

display(sorted_data)

We will define the `tokenizer` function takes a sentence as input, removes punctuation, converts the text to lowercase, tokenizes the sentence into words, removes English stopwords, and applies stemming using the Porter Stemmer. It helps in preparing text data for further natural language processing tasks like TF-IDF.

from nltk.tokenize import word_tokenize
from nltk.stem import PorterStemmer
from nltk.corpus import stopwords
import string

ENGLISH_STOP_WORDS = stopwords.words('english')
stemmer = PorterStemmer() 

def tokenizer(sentence):
    # remove punctuation and set to lower case
    for punctuation_mark in string.punctuation:
        sentence = sentence.replace(punctuation_mark,'').lower()

    # split sentence into words
    listofwords = sentence.split(' ')
    listofstemmed_words = []
    
    # remove stopwords and any tokens that are just empty strings
    for word in listofwords:
        if (not word in ENGLISH_STOP_WORDS) and (word!=''):
            # Stem words
            stemmed_word = stemmer.stem(word)
            listofstemmed_words.append(stemmed_word)

    return listofstemmed_words

Here, we're creating the TF-IDF (Term Frequency-Inverse Document Frequency) vectorizer, a crucial step in text feature extraction. We use the tokenizer function defined earlier to tokenize, preprocess, and transform the text data into a format suitable for TF-IDF computation. The `tokenizer` argument ensures that the tokenizer function we defined is used during the vectorization process.

# Create the TF-IDF vectorizer
tfidf_vectorizer = TfidfVectorizer(tokenizer=tokenizer, min_df=30, max_features=1000)

# Fit and transform the corpus using the vectorizer
tfidf_matrix = tfidf_vectorizer.fit_transform(sorted_data['features'])

# Print the shape of the TF-IDF matrix
print("TF-IDF Matrix Shape:", tfidf_matrix.shape)

# Create a DataFrame 'features' from the TF-IDF transformed data
features = pd.DataFrame(columns=tfidf_vectorizer.get_feature_names_out(), data=tfidf_matrix.toarray())

# Display the DataFrame
display(features)

The essential features we consider for similarity are the restaurant categories and their geographical locations. By computing the cosine similarity between these features, we can identify restaurants that share similar types of cuisines and are situated close to each other. This similarity-based approach allows us to provide relevant and personalized restaurant recommendations to users based on their preferences and the characteristics of restaurants they have enjoyed in the past. 

# Calculate the cosine similarity matrix 
cosine_similarity_matrix = cosine_similarity(features)

print("Shape of cosine_similarity_matrix:", cosine_similarity_matrix.shape)

The `restaurant_recommender` function suggests similar restaurants to a specified restaurant based on precomputed similarity scores. By filtering and extracting data for the given restaurant, it calculates similarity with other restaurants using a similarity matrix. The top 10 most similar restaurants are then recommended, providing users with personalized dining options that match their preferences and the characteristics of the chosen restaurant.

def restaurant_recommender(name, restaurants, similarities):
    # Get the restaurant by name
    restaurant_data = restaurants[restaurants['restaurant_name'] == name]
    
    # Extract the 'business_id' of the restaurant from the filtered data
    business_id = restaurant_data['business_id'].values[0]

    # Create a dataframe with the restaurant names and similarities
    sim_df = pd.DataFrame(
        {'restaurant': restaurants['restaurant_name'], 
         'similarity': similarities[business_id]
        })
    
    # Get the top 10 similar restaurants
    top_restaurants = sim_df.sort_values(by='similarity', ascending=False).head(10)
    
    return top_restaurants

# Test the recommender
similar_restaurants = restaurant_recommender("Wang's Shanghai Cuisine", sorted_data, cosine_similarity_matrix)
similar_restaurants.head(10)
